backend CRUD

+++ db/migrate/20220711023411_create_ingredients.rb
@@ -3,7 +3,7 @@ class CreateIngredients < ActiveRecord::Migration[6.1]
     create_table :ingredients do |t|
       t.string :name
       t.string :category
-      t.float :cost
+      t.integer :cost
       t.integer :points
 
       t.timestamps
